[RelEng] Fix tagging and PR creation in Release workflow
Previously the release-tag was added to the release commit (updating the
README), but as this PR is usually altered in the PR-submission process,
eventually the tag was off the linear history.
Now simply the current main branch head of the time when the workflow is
started is tagged, which is the commit before the release commit. That's
also the commit from which the release artifacts are build.

Also add missing permissions for the final PR creation step.